pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agentic-pipeline-ingestor"
version = "1.0.0"
description = "Enterprise-grade agentic data pipeline for document ingestion"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "API Support", email = "api-support@example.com"}
]
keywords = [
    "document-processing",
    "data-pipeline",
    "agentic",
    "llm",
    "ocr",
    "fastapi",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    # Web Framework
    "fastapi>=0.104.0,<1.0.0",
    "uvicorn[standard]>=0.24.0,<1.0.0",
    "python-multipart>=0.0.6",
    "starlette>=0.27.0",
    
    # Data Validation
    "pydantic>=2.5.0,<3.0.0",
    "pydantic-settings>=2.1.0",
    "email-validator>=2.1.0",
    
    # Database
    "sqlalchemy>=2.0.0,<3.0.0",
    "alembic>=1.12.0",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.9",
    
    # Caching & Queue
    "redis>=5.0.0",
    "hiredis>=2.2.0",
    
    # LLM Abstraction
    "litellm>=1.0.0",
    "openai>=1.0.0",
    
    # Document Processing
    "pymupdf>=1.23.0",
    "pillow>=10.0.0",
    "pdf2image>=1.16.0",
    "pytesseract>=0.3.10",
    "python-magic>=0.4.27; sys_platform != 'win32'",
    "python-magic-bin>=0.4.14; sys_platform == 'win32'",
    
    # HTTP Client
    "httpx>=0.25.0",
    "aiohttp>=3.9.0",
    
    # Observability
    "opentelemetry-api>=1.21.0",
    "opentelemetry-sdk>=1.21.0",
    "opentelemetry-instrumentation-fastapi>=0.42b0",
    "opentelemetry-instrumentation-sqlalchemy>=0.42b0",
    "opentelemetry-exporter-otlp>=1.21.0",
    "prometheus-client>=0.19.0",
    "structlog>=23.2.0",
    
    # Security
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-dotenv>=1.0.0",
    "cryptography>=41.0.0",
    
    # Utilities
    "orjson>=3.9.0",
    "PyYAML>=6.0.1",
    "click>=8.1.0",
    "tenacity>=8.2.0",
    "typing-extensions>=4.8.0",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "pytest-xdist>=3.3.0",
    "schemathesis>=3.21.0",
    "hypothesis>=6.88.0",
    "factory-boy>=3.3.0",
    "faker>=20.0.0",
    "respx>=0.20.0",
    "aiosqlite>=0.19.0",
    
    # Code Quality
    "ruff>=0.1.0",
    "mypy>=1.7.0",
    "pre-commit>=3.5.0",
    "bandit>=1.7.5",
    "safety>=2.3.0",
    
    # Documentation
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.4.0",
    "mkdocstrings[python]>=0.23.0",
]

docling = [
    "docling>=1.0.0",
]

azure = [
    "azure-identity>=1.15.0",
    "azure-storage-blob>=12.19.0",
    "azure-servicebus>=7.11.0",
    "azure-ai-vision>=0.13.0",
    "azure-monitor-opentelemetry>=1.1.0",
]

cognee = [
    "cognee-client>=0.1.0",
]

all = [
    "agentic-pipeline-ingestor[dev,docling,azure,cognee]",
]

[project.scripts]
pipeline-ingestor = "src.main:cli"

[project.urls]
Homepage = "https://github.com/example/agentic-pipeline-ingestor"
Documentation = "https://docs.pipeline.example.com"
Repository = "https://github.com/example/agentic-pipeline-ingestor"
Issues = "https://github.com/example/agentic-pipeline-ingestor/issues"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100
select = [
    "E",   # pycodestyle errors
    "F",   # Pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "W",   # pycodestyle warnings
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
    "TCH", # flake8-type-checking
    "TID", # flake8-tidy-imports
    "Q",   # flake8-quotes
    "FURB",# refurb
    "RUF", # Ruff-specific rules
]
ignore = [
    "E501",  # Line too long (handled by formatter)
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.isort]
known-first-party = ["src"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
implicit_reexport = false
strict_equality = true
show_error_codes = true

[[tool.mypy.overrides]]
module = [
    "litellm.*",
    "docling.*",
    "azure.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-fail-under=80",
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "contract: Contract tests",
    "slow: Slow tests",
]

[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/migrations/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
